README work
need to finish the README and writeup. After, I will finish tab 3. Plan is to add tab3 after test works, then move on to fix tab 1 because CONS doesn't have 'saving throw.' Also need to add another label that is the roll+modifier (which needs to account for if the box is checked and all that).