graph completed checkpoint
it seems like the graph is working how it should. Need to add the save button and possibly delete the interface.py and interfaceTest.py file because i think it was leftover testing folders without any use now. Also need to refactor the code once I move and rename the three tabtest files to the src folder.