Tab 3 added
tab 3 was fine in a window. Need to fix the geometry of the tabs. I could try and add a scroll bar, but that may be too much. Figure also saved with the probabilities. The python files for the tabtests were moved into src file (now I need to rename them)