renamed
renamed the files. TODO: Possibly add a scroll bar for only the attributes in tabone. Fix the proficiency so that it actually looks at the number and isn't stuck at 2. Move the labels over so that when clicked it isn't within the box. Fix CONS to have the saving throw. Finish the README and add the final write-up. Mess with the geometry to see if I can avoid a adding a scrollbar for tab 3. Create own gifs to add in place of the placeholders for the dice roller.