Adjuster label fixed
Now the modifier and proficiency bonus are properly added together. Next step is to add it to the dice roll. Probably need to save all of the values in a df (possibly) and the drop-down combo calls on that number to be added to the die roll. Other method could be to just dynamically look at the value in the label and add it (that might need the emit signal or something).