fixed some comments
going to go back to trying to fix the label. trying to get the signals was a bust. Might have to take the values of the modifier and proficiency and create labels for those at the bottom. So that there is a label for mod+die, mod+prof+die, and just die to see if that is any easier than getting the signal from the dropdowns to get the exact number with the roll