Dice roll calculation added
Had to work around the idea of an automatic calculation. The amount of values and signals to keep track of was much more than I had anticipated and kept resulting in errors that I was not able to fix. Even trying the method of storing them all in a df again was useless so there was a function added to calculate the roll. User puts in the number that they can see from the label in the attributes into a textbox and click calculate to add it to the dice roll. It is nice to see the initial roll to see if it was a fail or success. Comments and writeup also updated