Personal gif added
No code changed, I just wanted to add a gif I created for a die. This was a test to see if it worked and it did. Not 3d like the previous but I like it better and will try to to the other die that match this so it looks like they are all from the same set